remove php file modified js file
