--- views/default/pl/fe/matter/addressbook/edit.css
@@ -1,1 +1,0 @@
-p.form-control.active{background-color:#428bca;border-color:#428bca;color: #fff}
