+++ views/default/pl/fe/matter/addressbook/frame-loader.js
@@ -1,0 +1,8 @@
+require.config({
+  paths: {
+    "domReady": '/static/js/domReady',
+    "frame": '/views/default/pl/fe/matter/addressbook/frame',
+  },
+  urlArgs: "bust=" + (new Date()).getTime()
+});
+require(['frame']);
