+++ views/default/pl/fe/matter/addressbook/frame.css
@@ -1,0 +1,1 @@
+p.form-control.active{background-color:#428bca;border-color:#428bca;color: #fff}
