--- views/default/pl/fe/matter/addressbook/inapp.vw.php
@@ -1,7 +1,0 @@
-<?php
-include_once 'inmp.vw.php';
-
-//$view['params']['menu'] = '/rest/mp/app';
-$view['params']['layout-body'] = '/pl/fe/matter/inapp.tpl.htm';
-$view['params']['global_js'] = array('xxt.ui', true);
-$view['params']['angular-modules'] = "'ui.xxt'";
