--- views/default/pl/fe/matter/addressbook/setting.vw.php
@@ -1,2 +1,0 @@
-<?php
-$view['template'] = '/pl/fe/matter/addressbook/setting';
